blog_summarizer/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.0
beautifulsoup4==4.12.3
google-generativeai==0.8.0
python-dotenv==1.0.1
youtube-transcript-api==1.2.4
openai-whisper>=20231117
yt-dlp>=2024.1.0
